ci: fix if statements in ci.yaml
The if statements in ci.yaml were not working as expected. This commit
fixes them. The problem was that the if statements were referencing the
outputs of the changed-files job, but the changed-files job was not
actually running. This commit fixes that by referencing the outputs of
the changed-files job in the if statements of the build, lint, and test
jobs.